requirements.txt
Flask==2.3.3
google-api-python-client==2.94.0
google-auth==2.25.0
google-auth-oauthlib==1.1.0
gunicorn==21.2.0
